T701 修改无法删除当前添加的快捷操作； T702 修改无法修改资源类型的问题； T703 修改人员车辆无法正确绑定组的问题； T704 修改新建空组软件崩溃的问题； T705 修改删除终端后页面显示异常的问题； T706 修改无法更改用户累心的问题； T707 修改保存数据是软件卡顿的现象 T708 修改输入用户名按回车登录时提示未输入用户名的问题；
修改其他已知问题

根据新的配置文件协议，操作协议修改通信指令及流程（调度流程中增加SessionID）；
添加调度操作状态显示，包括开始，等待，异常等；

待处理：
配置用户的功能列表；
配置软件的功能列表；

+++ Dispatcher/modules/element.cs
@@ -12,7 +12,7 @@ namespace Dispatcher.Modules
 {
     public class CElement
     {
-        [DefaultValue((int)0), JsonProperty(PropertyName = "id")]
-        public int ID;
+        [DefaultValue((long)0), JsonProperty(PropertyName = "id")]
+        public long ID;
     }
 }
